[FEAT] New location of MOVE data #58 (#60)
Figured out the new location of the move data and it downloads now. Had to make some changes to readers.py and plotters.py, because the time decoding changed and wasn't monotonic anymore. Doesn't seem to clash with anything else though. Not sure why only one dataset is loaded, but this seems to be an issue with the rapid data as well so I will tackle this in a separate issue!